Close a cross-branch data leak, and two ways it could happen again
An aggressive staff-portal audit found three isolation defects. The first
was live and reproducible; the other two are the same mistake at earlier
stages of its life.

1. CROSS-BRANCH IDOR ON MEMBER TRAINING AND NUTRITION DATA (live)

A trainer with UserBranchAccess to Downtown only, asking about an Uptown
member:

  GET /api/members/{id}              -> 404   the member was hidden
  GET /api/workouts/logs/member/{id} -> 200   seven sessions, with loads

Also via workouts/assignments and nutrition/diet-plans. The member was
invisible; their training history was not.

WHY. Branch isolation is an EF global query filter, and a filter attaches
to an entity only if the entity carries the column it filters on. Member
is IBranchScoped. WorkoutLog, WorkoutAssignment, DietPlan and WaterLog are
tenant-scoped but carry NO BranchId -- verified in information_schema, not
inferred -- so four handlers written as `.Where(x => x.MemberId ==
request.MemberId)` were filtered by the caller's own route parameter and
by nothing else. Writes were never affected; they already validate through
the filtered Members set.

Every such read now goes through MemberScope.EnsureVisibleAsync, which
asks the already-filtered Members set whether the caller may see this
member at all -- borrowing the isolation that table has rather than
reimplementing branch logic per handler, which is what let four handlers
drift apart. NotFound, not Forbidden: "forbidden" would confirm the id is
real and turn the endpoint into an oracle for enumerating another branch's
membership, and 404 is what GET /api/members/{id} already returns.

Verified against the live API: all five endpoints now 404 for the foreign
member, and an own-branch member still returns their logs.

2. THE STAFF WATER LOG HAD NO UPPER BOUND

LogMyWaterCommand (the member's portal) capped at 5000ml. LogWaterCommand
-- which the portal DELEGATES INTO, and which staff call directly -- only
required "greater than zero". The strict rule guarded the weaker caller.

One int.MaxValue row overflowed GetNutritionReportQuery's int sum and
returned 500 for EVERY role, tenant-wide, until the row was deleted by
hand. Confirmed both directions: with the row present the report 500'd,
after deleting it, 200.

The bound now lives on the shared write path. The report also sums as
long, because a report is a read over whatever history already exists --
an aggregate that throws on its own data is the wrong failure.

3. THREE TRAINER TABLES HAD NO SCOPE COLUMNS AT ALL

TrainerSessions, TrainerAssignments and CommissionRecords carried neither
TenantId nor BranchId, so no filter could attach and handlers loaded rows
by raw id. Invisible on a single-tenant seed -- every test passes, every
screen looks right -- and a cross-tenant read AND WRITE on commission
money the day a second gym is onboarded. Same fix, same reason, as
20260811073831 applied to the workout and nutrition tables; these were
simply missed.

TenantId, not BranchId: a trainer belongs to a branch, but their
assignments and commission history follow the trainer, and the staff who
manage them are branch-mobile.

THE BACKFILL IS THE LOAD-BEARING PART. The tenant filter fails closed, so
adding the column alone would have hidden every historical row the moment
it deployed. Each row derives its tenant from the row it belongs to, in
dependency order. Verified after applying: 42 assignments, 3 commissions,
5 sessions, zero left on Guid.Empty.

And the seeder needed the same care: it runs from the CLI with no
authenticated user, so the ChangeTracker's tenant auto-stamp never fires
and a fresh seed would have produced Guid.Empty rows that the new filter
hides -- a gym seeded with no trainer assignments. It now sets TenantId
explicitly from the trainer, exactly as the WorkoutLog precedent does.
This is the part that would have bitten on the next reseed rather than in
a test.

771 tests green (7 new, including a negative control proving staff can
still read their OWN branch -- a guard like this fails by refusing
everyone, and a file that only asserts refusals would pass if it did).

+++ backend/src/GymOS.Application/Modules/Reports/Dtos/ReportDtos.cs
@@ -32,7 +32,8 @@ public record NutritionFoodItemRowDto(string FoodItemName, int TimesLogged, deci
 
 public record NutritionReportDto(
     List<NutritionFoodItemRowDto> TopFoodItems, int TotalMealEntriesLogged, decimal TotalCaloriesLogged,
-    int TotalWaterLogsLogged, int TotalWaterMlLogged);
+    // long, because this is a SUM over history and an int one overflowed into a tenant-wide 500.
+    int TotalWaterLogsLogged, long TotalWaterMlLogged);
 
 public record AtRiskMemberRowDto(
     Guid MemberId, string FullName, string MemberCode, DateOnly LastCheckInDate, int DaysSinceLastVisit);
